Backend/api_gateway/requirements.txt
fastapi
uvicorn[standard]
httpx
google-cloud-pubsub